fix(runtime): harden DSH session recovery
